extract specific clsuter providers to external calsses
