+++ app_build_suite/cluster_providers/__init__.py
@@ -1,0 +1,1 @@
+from .external_cluster_provider import ExternalClusterProvider  # noqa: F401
